+++ docs/conf.py
@@ -167,3 +167,19 @@ texinfo_documents = [
         "Miscellaneous",
     ),
 ]
+
+# This creates an 'Edit in GitLab' button on each documentation page.
+# This will propose a commit to the 'docs-staging' branch.
+# GitHub has an even better mechanism than GitLab, which gets the user
+# to propose a new branch with its own merge request, but unfortunately
+# we can't really achieve this without migrating to GitHub.
+#
+# See https://docs.readthedocs.io/en/stable/guides/edit-source-links-sphinx.html
+# for info on this specification.
+html_context = {
+    "display_gitlab": True, # Integrate Gitlab
+    "gitlab_user": "computational-audition-lab", # Username
+    "gitlab_repo": "psynet", # Repo name
+    "gitlab_version": "docs-staging", # Version
+    "conf_py_path": "/docs/", # Path in the checkout to the docs root
+}